requirements.txt
openai==1.25.1
langchain==0.1.20
langchain-openai==0.1.7
langchain-community==0.0.38
python-dotenv==1.0.1
faiss-cpu==1.8.0
pypdf==4.1.0
requests==2.31.0
pydantic==2.7.1
pydantic-settings==2.2.1
uvicorn==0.29.0
fastapi==0.111.0
httpx==0.27.0
langchainhub==0.1.15
